Refactor : UserType 을 enums.type으로 이동시켰습니다.
