+++ src/main/java/shootingstar/var/enums/type/UserType.java
@@ -1,4 +1,4 @@
-package shootingstar.var.entity;
+package shootingstar.var.enums.type;
 
 public enum UserType {
     ROLE_ADMIN, ROLE_VIP, ROLE_BASIC
